+++ fs/test/test.cpp
@@ -145,8 +145,8 @@ TEST(Path, level_valid_ness)
 
 TEST(string_view, equality) // dependented by `Path`
 {
-    fs::string_view a(nullptr, 0UL);
-    fs::string_view b((char*)234, 0UL);
+    fs::string_view a(nullptr, 0ULL);
+    fs::string_view b((char*)234, 0ULL);
     EXPECT_EQ(a, b);
 }
 
@@ -165,12 +165,12 @@ TEST(Tree, node)
     node.creat(k1234, (void*)1234);
     node.creat(k1234, (void*)2345);
     node.creat(k1234, v1234);
-    EXPECT_EQ(node.size(), 5ul);
+    EXPECT_EQ(node.size(), 5ull);
 
     for (auto x: subnodes)
         node.mkdir(x);
 
-    EXPECT_EQ(node.size(), 9ul);
+    EXPECT_EQ(node.size(), 9ull);
 
     f = F;
     void* v;
@@ -625,9 +625,9 @@ class ExampleAsyncDir: public AsyncDIR {
     }
     OVERRIDE_ASYNC0(void, rewinddir) {
     }
-    OVERRIDE_ASYNC(void, seekdir, long loc) {
+    OVERRIDE_ASYNC(void, seekdir, long long loc) {
     }
-    OVERRIDE_ASYNC0(long, telldir){
+    OVERRIDE_ASYNC0(long long, telldir){
     }
 };
 
@@ -751,8 +751,8 @@ TEST(range_split, sub_range)
     EXPECT_FALSE(sr);
     sr.assign(0, 233, 1024);
     EXPECT_TRUE(sr);
-    EXPECT_EQ(233ul, sr.begin());
-    EXPECT_EQ(233ul+1024ul, sr.end());
+    EXPECT_EQ(233ull, sr.begin());
+    EXPECT_EQ(233ull+1024ull, sr.end());
     sr.clear();
     EXPECT_FALSE(sr);
     sr.assign(1, 233, 1024);
@@ -765,25 +765,25 @@ TEST(range_split, range_split_simple_case)
     // with abegin, aend as 0, 11
     // 11 parts in total
     EXPECT_FALSE(split.small_note);
-    EXPECT_EQ(42ul, split.begin);
-    EXPECT_EQ(363ul, split.end);
-    EXPECT_EQ(1ul, split.abegin);
-    EXPECT_EQ(12ul, split.aend);
-    EXPECT_EQ(2ul, split.apbegin);
-    EXPECT_EQ(11ul, split.apend);
-    EXPECT_EQ(32ul, split.aligned_begin_offset());
-    EXPECT_EQ(384ul, split.aligned_end_offset());
+    EXPECT_EQ(42ull, split.begin);
+    EXPECT_EQ(363ull, split.end);
+    EXPECT_EQ(1ull, split.abegin);
+    EXPECT_EQ(12ull, split.aend);
+    EXPECT_EQ(2ull, split.apbegin);
+    EXPECT_EQ(11ull, split.apend);
+    EXPECT_EQ(32ull, split.aligned_begin_offset());
+    EXPECT_EQ(384ull, split.aligned_end_offset());
     auto p = split.all_parts();
-    EXPECT_EQ(1ul, p.begin()->i);
-    EXPECT_EQ(10ul, p.begin()->begin());
-    EXPECT_EQ(32ul, p.begin()->end());
-    EXPECT_EQ(12ul, p.end()->i);
+    EXPECT_EQ(1ull, p.begin()->i);
+    EXPECT_EQ(10ull, p.begin()->begin());
+    EXPECT_EQ(32ull, p.begin()->end());
+    EXPECT_EQ(12ull, p.end()->i);
     uint64_t cnt = 1;
     for (auto &rs: p) {
         EXPECT_EQ(cnt++, rs.i);
         if (rs != p.begin() && rs != p.end()) {
-            EXPECT_EQ(0ul, rs.begin());
-            EXPECT_EQ(32ul, rs.end());
+            EXPECT_EQ(0ull, rs.begin());
+            EXPECT_EQ(32ull, rs.end());
         }
     }
     split = fs::range_split(2, 12, 24);
@@ -801,27 +801,27 @@ TEST(range_split, range_split_aligned_case)
     // it should be split into [begin, end) as [32, 64)+[64, 76) +... +[352,353)
     // with abegin, aend as 0, 11
     // 11 parts in total
-    EXPECT_EQ(32ul, split.begin);
-    EXPECT_EQ(353ul, split.end);
-    EXPECT_EQ(1ul, split.abegin);
-    EXPECT_EQ(12ul, split.aend);
-    EXPECT_EQ(1ul, split.apbegin);
-    EXPECT_EQ(11ul, split.apend);
+    EXPECT_EQ(32ull, split.begin);
+    EXPECT_EQ(353ull, split.end);
+    EXPECT_EQ(1ull, split.abegin);
+    EXPECT_EQ(12ull, split.aend);
+    EXPECT_EQ(1ull, split.apbegin);
+    EXPECT_EQ(11ull, split.apend);
     auto p = split.all_parts();
     EXPECT_FALSE(split.is_aligned());
     EXPECT_TRUE(split.is_aligned(128));
     EXPECT_TRUE(split.is_aligned_ptr((const void*)(uint64_t(65536))));
-    EXPECT_EQ(1ul, p.begin()->i);
-    EXPECT_EQ(0ul, p.begin()->begin());
-    EXPECT_EQ(32ul, p.begin()->end());
-    EXPECT_EQ(12ul, p.end()->i);
-    EXPECT_EQ(352ul, split.aligned_length());
+    EXPECT_EQ(1ull, p.begin()->i);
+    EXPECT_EQ(0ull, p.begin()->begin());
+    EXPECT_EQ(32ull, p.begin()->end());
+    EXPECT_EQ(12ull, p.end()->i);
+    EXPECT_EQ(352ull, split.aligned_length());
     auto q = split.aligned_parts();
     uint64_t cnt = 1;
     for (auto &rs: q) {
         EXPECT_EQ(cnt++, rs.i);
-        EXPECT_EQ(0ul, rs.begin());
-        EXPECT_EQ(32ul, rs.end());
+        EXPECT_EQ(0ull, rs.begin());
+        EXPECT_EQ(32ull, rs.end());
     }
     split = fs::range_split(0, 23, 24);
     EXPECT_TRUE(split.postface);
@@ -850,32 +850,32 @@ TEST(range_split_power2, basic) {
         LOG_DEBUG(rs.i, ' ', rs.begin(), ' ', rs.end());
     }
     EXPECT_FALSE(split.small_note);
-    EXPECT_EQ(42ul, split.begin);
-    EXPECT_EQ(363ul, split.end);
-    EXPECT_EQ(1ul, split.abegin);
-    EXPECT_EQ(12ul, split.aend);
-    EXPECT_EQ(2ul, split.apbegin);
-    EXPECT_EQ(11ul, split.apend);
-    EXPECT_EQ(32ul, split.aligned_begin_offset());
-    EXPECT_EQ(384ul, split.aligned_end_offset());
+    EXPECT_EQ(42ull, split.begin);
+    EXPECT_EQ(363ull, split.end);
+    EXPECT_EQ(1ull, split.abegin);
+    EXPECT_EQ(12ull, split.aend);
+    EXPECT_EQ(2ull, split.apbegin);
+    EXPECT_EQ(11ull, split.apend);
+    EXPECT_EQ(32ull, split.aligned_begin_offset());
+    EXPECT_EQ(384ull, split.aligned_end_offset());
     auto p = split.all_parts();
-    EXPECT_EQ(p.begin()->i, 1ul);
-    EXPECT_EQ(p.begin()->begin(), 10ul);
-    EXPECT_EQ(p.begin()->end(), 32ul);
-    EXPECT_EQ(p.end()->i, 12ul);
+    EXPECT_EQ(p.begin()->i, 1ull);
+    EXPECT_EQ(p.begin()->begin(), 10ull);
+    EXPECT_EQ(p.begin()->end(), 32ull);
+    EXPECT_EQ(p.end()->i, 12ull);
     uint64_t cnt = 1;
     for (auto &rs: p) {
         EXPECT_EQ(rs.i, cnt++);
         if (rs != p.begin() && rs != p.end()) {
-            EXPECT_EQ(rs.begin(), 0ul);
-            EXPECT_EQ(rs.end(), 32ul);
+            EXPECT_EQ(rs.begin(), 0ull);
+            EXPECT_EQ(rs.end(), 32ull);
         }
     }
 }
 
 TEST(range_split_power2, random_test) {
     uint64_t offset = rand(), length = rand();
-    uint64_t interval = 1UL << (rand()%32 + 1);
+    uint64_t interval = 1ULL<< (rand()%32 + 1);
     fs::range_split_power2 split(offset, length, interval);
     EXPECT_EQ(offset, split.begin);
     EXPECT_EQ(offset + length, split.end);
@@ -891,15 +891,15 @@ TEST(range_split_vi, basic) {
     uint64_t kp[] = {0, 32, 64, 128, 256, 512, UINT64_MAX};
     fs::range_split_vi split(12, 321, kp, 7);
     uint64_t *it = kp;
-    EXPECT_EQ(12ul, split.begin);
-    EXPECT_EQ(333ul, split.end);
-    EXPECT_TRUE(split.is_aligned(0ul));
-    EXPECT_FALSE(split.is_aligned(1ul));
-    EXPECT_TRUE(split.is_aligned(128ul));
+    EXPECT_EQ(12ull, split.begin);
+    EXPECT_EQ(333ull, split.end);
+    EXPECT_TRUE(split.is_aligned(0ull));
+    EXPECT_FALSE(split.is_aligned(1ull));
+    EXPECT_TRUE(split.is_aligned(128ull));
     for (auto p : split.all_parts()) {
         LOG_DEBUG(p.i, ' ', p.begin(), ' ', p.end());
-        EXPECT_EQ(*it == 0ul ? 12ul : 0ul, p.begin());
-        EXPECT_EQ(*it == 256ul ? 321ul-256ul+12ul : (*(it+1) - *it), p.end());
+        EXPECT_EQ(*it == 0ull ? 12ull : 0ull, p.begin());
+        EXPECT_EQ(*it == 256ull ? 321ull-256ull+12ull : (*(it+1) - *it), p.end());
         it++;
     }
     uint64_t kpfail[] = {0, 32, 796, 128, 256, 512, UINT64_MAX};
@@ -916,14 +916,14 @@ TEST(range_split_vi, left_side_aligned) {
     uint64_t kp[] = {0, 32, 64, 128, 256, 512, UINT64_MAX};
     fs::range_split_vi split(0, 256, kp, 7);
     uint64_t *it = kp;
-    EXPECT_EQ(0ul, split.begin);
-    EXPECT_EQ(256ul, split.end);
-    EXPECT_TRUE(split.is_aligned(0ul));
-    EXPECT_FALSE(split.is_aligned(1ul));
-    EXPECT_TRUE(split.is_aligned(128ul));
+    EXPECT_EQ(0ull, split.begin);
+    EXPECT_EQ(256ull, split.end);
+    EXPECT_TRUE(split.is_aligned(0ull));
+    EXPECT_FALSE(split.is_aligned(1ull));
+    EXPECT_TRUE(split.is_aligned(128ull));
     for (auto p : split.all_parts()) {
         LOG_DEBUG(p.i, ' ', p.begin(), ' ', p.end());
-        EXPECT_EQ(0ul, p.begin());
+        EXPECT_EQ(0ull, p.begin());
         EXPECT_EQ((*(it+1) - *it), p.end());
         it++;
     }
